+++ src/allotropy/parsers/beckman_vi_cell_blu/vi_cell_blu_reader.py
@@ -1,3 +1,5 @@
+import unicodedata
+
 import pandas as pd
 
 from allotropy.allotrope.pandas_util import read_csv
@@ -17,4 +19,14 @@ class ViCellBluReader:
         named_file_contents.contents.seek(0)
         return read_csv(
             named_file_contents.contents, index_col=False, encoding=encoding
-        )
+        ).rename(columns=cls._normalize_column_names)
+
+    @classmethod
+    def _normalize_column_names(cls, col: str) -> str:
+        """Normalize column names that look the same but have mismatching characters
+
+        Some column names in the instrument files include two different `µ` characters that look the same,
+        the Micro Sign (U+00B5) and the Greek Small letter mu (U+03BC), this mapper uses the unicode `NFKC`
+        normalization to ensure consistency when retrieving values from those columns.
+        """
+        return unicodedata.normalize("NFKC", col)